pyproject.toml
[tool.poetry]
name = "companion-backend"
version = "0.1.0"
description = ""
authors = ["duggalr <duggalr42@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.13"
poetry-dotenv = "^0.4.0"
fastapi = {extras = ["standard"], version = "^0.115.5"}
openai = "^1.54.4"
docker = "^7.1.0"
celery = "^5.4.0"
sqlalchemy = "^2.0.36"
psycopg2-binary = "^2.9.10"
alembic = "^1.14.0"
redis = "^5.2.0"
uvicorn = "^0.32.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
